* add .vscode/ custom workspace settings to .gitignore
* ensure consistent single blank line between methods and single final newline

* remove unnecessary comments in app.component.html

* move interfaces to its own *.model.ts files

* change various variables to conform to camel case
