+++ templates/Web/Pages/Angular.Grid/src/app/app-shell/Param_SourceName_Kebab/grid.component.ts
@@ -1,6 +1,7 @@
 ﻿import { Component, OnInit } from '@angular/core';
 
-import { GridService, IGridTextItem } from './grid.service';
+import { GridService } from './grid.service';
+import { IGridTextItem } from './grid.model';
 
 @Component({
   selector: 'app-grid',
@@ -9,9 +10,9 @@ import { GridService, IGridTextItem } from './grid.service';
 })
 export class GridComponent implements OnInit {
 
-  GreyBox = require('../../../assets/GreyBox.svg') as string;
-  WarningMessageText = 'Request to get grid text failed:';
-  WarningMessageOpen = false;
+  greyBox = require('../../../assets/GreyBox.svg') as string;
+  warningMessageText = 'Request to get grid text failed:';
+  warningMessageOpen = false;
   gridTextAssets: IGridTextItem[] = [
     {
       description: 'example1',
@@ -24,21 +25,23 @@ export class GridComponent implements OnInit {
       id: 1
     }
   ];
+
   constructor(private gridService: GridService) { }
 
   ngOnInit() {
     this.gridService.getGridItems().subscribe(
-      result => {
+      (result: IGridTextItem[]) => {
         this.gridTextAssets = result;
       },
       error => {
-        this.WarningMessageOpen = true;
-        this.WarningMessageText = `Request to get grid text failed: ${error}`;
+        this.warningMessageOpen = true;
+        this.warningMessageText = `Request to get grid text failed: ${error}`;
       }
     );
   }
+
   handleWarningClose(open: boolean) {
-    this.WarningMessageOpen = open;
-    this.WarningMessageText = '';
+    this.warningMessageOpen = open;
+    this.warningMessageText = '';
   }
 }
