+++ templates/Web/Pages/Angular.Grid/src/app/app-shell/Param_SourceName_Kebab/grid.model.ts
@@ -1,0 +1,5 @@
+export interface IGridTextItem {
+  description: string;
+  header: string;
+  id: number;
+}
