+++ templates/Web/Pages/Angular.List/src/app/app-shell/Param_SourceName_Kebab/list.component.html
@@ -16,9 +16,9 @@
     </app-list-item>
 
     <app-warning-message
-      *ngIf="WarningMessageOpen"
-      [text]="WarningMessageText"
-      (WarningMessageOpen)="handleWarningClose($event)"
+      *ngIf="warningMessageOpen"
+      [text]="warningMessageText"
+      (warningMessageOpen)="handleWarningClose($event)"
     ></app-warning-message>
   </div>
 </main>
