+++ templates/Web/Pages/Angular.List/src/app/app-shell/Param_SourceName_Kebab/list.model.ts
@@ -1,0 +1,4 @@
+export interface IListItem {
+  _id: number;
+  text: string;
+}
