+++ templates/Web/Pages/Angular.MasterDetail/src/app/app-shell/Param_SourceName_Kebab/master-detail-page/master-detail-page.component.ts
@@ -1,5 +1,5 @@
 ﻿import { Component, OnInit, Input } from '@angular/core';
-import { IMasterDetailText } from '../master-detail.service';
+import { IMasterDetailText } from '../master-detail.model';
 
 @Component({
   selector: 'app-master-detail-page',
@@ -9,6 +9,7 @@ import { IMasterDetailText } from '../master-detail.service';
 export class MasterDetailPageComponent implements OnInit {
 
   @Input() textSampleData: IMasterDetailText;
+
   constructor() { }
 
   ngOnInit() {
@@ -23,5 +24,4 @@ export class MasterDetailPageComponent implements OnInit {
       id: 0
     };
   }
-
 }
