+++ templates/Web/Pages/Angular.MasterDetail/src/app/app-shell/Param_SourceName_Kebab/master-detail.model.ts
@@ -1,0 +1,10 @@
+export interface IMasterDetailText {
+  title: string;
+  id: number;
+  status: string;
+  orderDate: string;
+  shipTo: string;
+  orderTotal: number;
+  shortDescription: string;
+  longDescription: string;
+}
