+++ templates/Web/Projects/AngularDefault/src/app/app.component.html
@@ -1,11 +1,5 @@
-﻿<!--The content below is only a placeholder and can be replaced.-->
-
-<!-- header -->
-<app-nav-bar></app-nav-bar>
+﻿<app-nav-bar></app-nav-bar>
 
 <router-outlet></router-outlet>
 
-<!-- footer -->
 <app-footer></app-footer>
-
-
