+++ templates/Web/Projects/AngularDefault/src/app/app.component.ts
@@ -6,5 +6,6 @@
   styleUrls: ['./app.component.css']
 })
 export class AppComponent {
+
   title = 'wts.AngularDefault';
 }
